fix: ID's unified to long, sql scripts fixed, controller updated for usage long ids

+++ src/main/java/com/nesterrovv/vpn/subscription/controller/SubscriptionController.java
@@ -82,7 +82,7 @@ public class SubscriptionController {
         Token oldToken = current.getToken();
         if (oldToken != null) {
             List<Subscription> subscriptionsWithThisToken =
-                subscriptionService.getSubscriptionsByLinkedTokenId(oldToken.getId());
+                subscriptionService.getSubscriptionsByLinkedTokenId((long) oldToken.getId());
             if (!subscriptionsWithThisToken.isEmpty()) {
                 subscriptionService.updateSubscription(id, expirationDate, false, null);
                 tokenService.delete(oldToken.getId());
